servidor/requirements.txt
requests
pandas
numpy
paho-mqtt